requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
protobuf==4.21.12
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
